api/requirements.txt
openai==1.35.5
langchain==0.2.6
langchain-core==0.2.10
langchain-text-splitters==0.2.2
langsmith==0.1.82
flask==3.0.3
langchain-openai==0.1.10
langchain-community==0.2.6
load-dotenv==0.1.0
flask-cors==4.0.1
beautifulsoup4==4.12.3
pymongo==4.8.0
PyPDF2==3.0.1
pypdf==4.3.0
unstructured==0.15.0
fastapi==0.112.0
uvicorn==0.30.5
gunicorn==23.0.0